Placing team mutations
