+++ src/Mutations/profileMutation.tsx
@@ -1,0 +1,38 @@
+import { gql } from '@apollo/client';
+
+export const UPDATE_PROFILE = gql`
+  mutation UpdateProfile(
+    $lastName: String
+    $firstName: String
+    $address: String
+    $city: String
+    $country: String
+    $phoneNumber: String
+    $biography: String
+    $fileName: String
+    $cover: String
+    $githubUsername: String
+  ) {
+    updateProfile(
+      lastName: $lastName
+      firstName: $firstName
+      address: $address
+      city: $city
+      country: $country
+      phoneNumber: $phoneNumber
+      biography: $biography
+      fileName: $fileName
+      cover: $cover
+      githubUsername: $githubUsername
+    ) {
+      id
+      lastName
+      firstName
+      biography
+      phoneNumber
+      address
+      city
+      country
+    }
+  }
+`;
